Updated unit test for the summary module

+++ tests/test_summary.py
@@ -70,7 +70,7 @@ class TestSummary(unittest.TestCase):
                                   []]
             })
         # Set the data type of the "ld" column to int64
-        expected_result['ld'] = expected_result['ld'].astype('int32')
+        #expected_result['ld'] = expected_result['ld'].astype('int32')
 
         try:
             pd.testing.assert_frame_equal(expected_result, actual_result)
